Create RubiksSimulator.py
Began writing the class for a 3x3x3 Rubik's cube called Rubiks27. It is initialized to be solved. The cube is internally represented as a dictionary. Began but have not finished the front_clockwise method which should simulate turning the front face clockwise and correctly update all of the cells of the cube.